requirements.txt
langchain
openai
cohere
huggingface_hub
python-dotenv
streamlit